+++ src/cpu/instructions/adc/mod.rs
@@ -1,5 +1,2 @@
 pub mod decoder;
 pub mod opcode;
-
-mod adc;
-mod add;
