+++ src/cpu/instructions/mod.rs
@@ -1,0 +1,9 @@
+pub mod adc;
+pub mod add;
+pub mod decoder;
+pub mod opcodes;
+pub mod operand;
+pub mod instructions;
+
+mod opcode;
+mod test;
