--- src/cpu/opcodes/mod.rs
@@ -1,7 +1,0 @@
-pub mod add;
-pub mod adc;
-pub mod decoders;
-pub mod operand;
-
-mod opcode;
-mod test_util;
